pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yogu-reader-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-adk>=1.9.0",
    "langchain>=0.3.27",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.35.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.0.3,<1.1.0",
    "asyncio",
    "litellm>=1.74.9.post1",
    "arrow>=1.3.0",
    "langchain-tavily>=0.2.11",
    "celery[redis]>=5.5.3",
    "pymupdf>=1.23.0",
    "pymilvus[model]>=2.5.8",
    "milvus-model>=0.2.12",
    "minio==7.2.16",
    "dashscope>=1.24.1",
    "langchain-openai>=0.3.28",
    "langchain-core>=0.3.74",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.uv]
index-url = "https://nexus.yogu.team/repository/pypi/simple"
